Day 7 - bags containing bags
this one was tricky. Some more figuring out of the finer points of data
structures and iteration in javascript. For example the fact that `Object.keys`
is a static function rather than a member function of the Object which might be
invoked `{}.keys()` which had me puzzled/stumped for a while. There appears to be a `Map` class in javascript standard library which does have a `keys` method.